moved friends list on right and add tabs
