+++ app/components/shared/AppNavigation.tsx
@@ -15,28 +15,8 @@ interface AppNavigationProps {
 const AppNavigation = ({ isAuthenticated }: AppNavigationProps) => (
   <Navbar isBordered className="top-navbar">
     <NavbarBrand>
-      <p className="font-bold text-inherit">UserBook</p>
+      <p className="font-bold text-inherit">Friends Book</p>
     </NavbarBrand>
-
-    {/* {isAuthenticated && (
-      <NavbarContent className="hidden sm:flex gap-4" justify="center">
-        <NavbarItem>
-          <Link color="foreground" href="#">
-            Features
-          </Link>
-        </NavbarItem>
-        <NavbarItem isActive>
-          <Link href="#" aria-current="page">
-            Customers
-          </Link>
-        </NavbarItem>
-        <NavbarItem>
-          <Link color="foreground" href="#">
-            Integrations
-          </Link>
-        </NavbarItem>
-      </NavbarContent>
-    )} */}
 
     <NavbarContent justify="end">
       {isAuthenticated ? (
@@ -48,16 +28,11 @@ const AppNavigation = ({ isAuthenticated }: AppNavigationProps) => (
           </Form>
         </NavbarItem>
       ) : (
-        <>
-          {/* <NavbarItem className="hidden lg:flex">
-            <Link href="#">Login</Link>
-          </NavbarItem> */}
-          <NavbarItem>
-            <Button as={Link} color="primary" href="#" variant="flat">
-              Sign Up
-            </Button>
-          </NavbarItem>
-        </>
+        <NavbarItem>
+          <Button as={Link} color="primary" href="#" variant="flat">
+            Sign Up
+          </Button>
+        </NavbarItem>
       )}
     </NavbarContent>
   </Navbar>
